# Fuse write kv buffer into rope for qwen3 moe & bailing moe