@b8zhong 2025-09-22
@yuan-luo Q: do you know why this is only applicable to BF16 cache and MoE? Or can it be adapted to dense models.